requirements.txt
pandas
google-cloud-bigquery
pandas-gbq
datetime
pytz
google_play_scraper
app_store_scraper
findspark
pyspark
pyarrow